# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M-PESA Business Loans API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: Get the current customer's profile
      responses:
        '200':
          description: Profile found.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CustomerProfile'
                  - type: 'null'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      tags: [Profile]
      operationId: updateMyProfile
      summary: Create or update the current customer's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerProfileInput'
      responses:
        '200':
          description: Profile saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProfile'
  /profile/documents:
    get:
      tags: [Profile]
      operationId: listMyDocuments
      summary: List documents uploaded by the current customer
      responses:
        '200':
          description: List of documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      tags: [Profile]
      operationId: addMyDocument
      summary: Record an uploaded document (ID front/back, selfie, supporting)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentInput'
      responses:
        '201':
          description: Document recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
  /profile/otp/request:
    post:
      tags: [Profile]
      operationId: requestPhoneOtp
      summary: Generate an OTP code for phone verification (stubbed delivery)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequestInput'
      responses:
        '200':
          description: OTP generated. In this foundation phase, the code is returned directly since no SMS provider is connected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpRequestResult'
  /profile/otp/verify:
    post:
      tags: [Profile]
      operationId: verifyPhoneOtp
      summary: Verify a phone OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyInput'
      responses:
        '200':
          description: Verification result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpVerifyResult'
        '400':
          description: Invalid or expired code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # ---------- Loans (customer-facing) ----------
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ----- Storage -----
...
    CustomerProfile:
      type: object
      required:
        - id
        - userId
        - phone
        - phoneVerified
        - dateOfBirth
        - address
        - city
        - nationalIdNumber
        - idFrontUrl
        - idBackUrl
        - selfieUrl
        - profileComplete
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        userId: { type: string }
        phone: { type: [string, 'null'] }
        phoneVerified: { type: boolean }
        dateOfBirth: { type: [string, 'null'] }
        address: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        nationalIdNumber: { type: [string, 'null'] }
        idFrontUrl: { type: [string, 'null'] }
        idBackUrl: { type: [string, 'null'] }
        selfieUrl: { type: [string, 'null'] }
        profileComplete: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CustomerProfileInput:
      type: object
      properties:
        phone: { type: string }
        dateOfBirth: { type: string }
        address: { type: string }
        city: { type: string }
        nationalIdNumber: { type: string }
        idFrontUrl: { type: string }
        idBackUrl: { type: string }
        selfieUrl: { type: string }
    DocumentType:
      type: string
      enum: [id_front, id_back, selfie, supporting]
    Document:
      type: object
      required: [id, customerId, type, fileName, fileUrl, uploadedAt]
      properties:
        id: { type: string }
        customerId: { type: string }
        type:
          $ref: '#/components/schemas/DocumentType'
        fileName: { type: string }
        fileUrl: { type: string }
        uploadedAt: { type: string, format: date-time }
    DocumentInput:
      type: object
      required: [type, fileName, fileUrl]
      properties:
        type:
          $ref: '#/components/schemas/DocumentType'
        fileName: { type: string, minLength: 1 }
        fileUrl: { type: string, minLength: 1 }
    OtpRequestInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 1
    OtpRequestResult:
      type: object
      required: [message, devCode]
      properties:
        message:
          type: string
        devCode:
          type: string
          description: The generated OTP, returned directly since no SMS provider is connected yet (foundation phase stub).
    OtpVerifyInput:
      type: object
      required: [phone, code]
      properties:
        phone: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
    OtpVerifyResult:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean

    # ----- Loans -----
...